Over engineered transition method
Added a super complex and possibly unnecessary means to transition
frames. Wanted to make it such that each frame would not know of each
others existence.

I WOULD NOT RECOMMEND DOING THIS EVER: However, I did learn a lot doing
this and it was worth the 5 hours I spent learning the event sytstem of
wxWidgets

TLDR:
Router holds dependencies for all custom frames, it acts as a backbone
event handler that listens for a custom event type called
TransitionEvent that is kicked up from the spawning frames that the
router emits by passing a generic form of itself to each instance of the
frame it spawns.

This was stupid